[NEO]: Fix potential OOB
* Fix lint issues

+++ src/NEO/Transaction.cpp
@@ -14,11 +14,11 @@ using namespace TW;
 
 namespace TW::NEO {
 
-int64_t Transaction::size() const {
+size_t Transaction::size() const {
     return serialize().size();
 }
 
-void Transaction::deserialize(const Data& data, int initial_pos) {
+void Transaction::deserialize(const Data& data, size_t initial_pos) {
     type = (TransactionType)data[initial_pos++];
     version = data[initial_pos++];
     initial_pos = deserializeExclusiveData(data, initial_pos);
@@ -30,7 +30,7 @@ void Transaction::deserialize(const Data& data, int initial_pos) {
     Serializable::deserialize<TransactionOutput>(outputs, data, initial_pos);
 }
 
-Transaction* Transaction::deserializeFrom(const Data& data, int initial_pos) {
+Transaction* Transaction::deserializeFrom(const Data& data, size_t initial_pos) {
     Transaction* resp = nullptr;
     switch ((TransactionType)data[initial_pos]) {
     case TransactionType::TT_MinerTransaction:
